add 1 to `previous_block_ip` to avoid that all the recursive basic blocks map to 0
